Adding of user-defined type conversions for python bindings inside module directories Adding of destructor and placement new constructors for classes wrapped with CV_EXPORTS_W_SIMPLE macro
