+++ modules/features2d/misc/python/pyopencv_features2d.hpp
@@ -1,0 +1,3 @@
+#ifdef HAVE_OPENCV_FEATURES2D
+typedef SimpleBlobDetector::Params SimpleBlobDetector_Params;
+#endif
