+++ modules/stitching/misc/python/pyopencv_stitching.hpp
@@ -1,0 +1,9 @@
+#ifdef HAVE_OPENCV_STITCHING
+typedef Stitcher::Status Status;
+
+template<>
+PyObject* pyopencv_from(const Status& value)
+{
+    return PyInt_FromLong(value);
+}
+#endif
